changing skip and limit params in Index and Plans to make it long instead of int

+++ src/foam/dao/index/Index.java
@@ -21,7 +21,7 @@ public interface Index {
   public Object remove(Object state, FObject value);
   public Object removeAll();
   public FindPlan planFind(Object state, Object key);
-  public SelectPlan planSelect(Object state, Sink sink, int skip, int limit, Comparator order, Predicate predicate);
+  public SelectPlan planSelect(Object state, Sink sink, long skip, long limit, Comparator order, Predicate predicate);
   public long size(Object state);
 
   // Future:
